Record condition for charts from Helm repo
To be able to set conditions reporting the status of charts from Helm
repos, the bit doing the condition setting must know what happens when
attempting to fetch the charts. That bit is the `chartsync` package,
which has the core loop reconciling the state of the cluster, the git
mirroring, and (now) the cached charts. So, move the chart fetching
code, and its call sites, to `chartsync`.

Moving the chart fetching code from `release` to `chartsync` makes
sense for another reason: it means the argument to
`release.Install(..)` can be a path to the chart, uniformly, and no
interpretation of the chart source in FluxHelmRelease is required
there.